requirements.txt
# Lighthouse Bridge — Python Dependencies
Flask>=3.0.0
PyYAML>=6.0
requests>=2.31.0
mysql-connector-python>=9.0.0
gunicorn>=22.0.0
